+++ CHANGELOG.md
@@ -1,5 +1,24 @@
 # es6/7-sass-webpack-starter changelog
 
+## 1.1.3
+
+* This is a bone-head patch because I forgot to commit CHANGELOG.md
+
+## 1.1.2
+
+* added support for JsDoc generated documentation
+* added not-supported.html page and module to handle UAs with poor feature support
+
+## 1.1.1
+
+* resolved webpack config relative path issue when in dev mode
+* thanks to [@mastilver](https://github.com/jantimon/html-webpack-plugin/issues/665#issuecomment-326794429) on the html-webpack-plugin team
+
+## 1.1.0
+
+* added `<head>` template guidance
+* added not-found.html page and module to handle 404s
+
 ## 1.0.4
 
 * reorg of src/